Refactor: 별도의 컨트롤러와 서비스로 분리함
